refactor: 重构 TypeScript 项目结构
- 更新 .gitignore 文件，排除新的 node_modules 目录
- 新增 nacos_mcp_router_typescript 目录，用于 TypeScript 项目
- 移动 TypeScript 相关文件到新目录下
- 更新 TypeScript 文件中的导入路径
- 重命名 package-lock.json 和 package.json 文件
