+++ src/nacos_mcp_router_typescript/src/mcp_manager.ts
