+++ src/nacos_mcp_router_typescript/src/router.ts
