+++ src/nacos_mcp_router_typescript/src/tool_manager.ts
