+++ src/nacos_mcp_router_typescript/tsconfig.json
